Add pagination
For seeing all search results multiple pages are needed to display the
hits. This needs support in the search functions, they need to be able to start
from a specified search result on and they also need to adjust the number of
returned hits. Also there are some additional state Variables needed.

The navigation is done by buttons at the end of the page. There is also a
possibility to change how many results are displayed on one page.